# lib/api-spec/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Api",
    "version": "0.2.0",
    "description": "Single-user persisted wardrobe API. Generated from shared Zod schemas."
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
...
  "paths": {
...
    "/wardrobe": {
      "get": {
        "operationId": "listClosetItems",
        "tags": [
          "closet"
        ],
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "Tops",
                "Bottoms",
                "Outerwear",
                "Dresses",
                "Shoes",
                "Bags",
                "Accessories",
                "Jewelry"
              ]
            }
          },
          {
            "name": "favorite",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            }
          },
          {
            "name": "color",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "season",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "Spring",
                "Summer",
                "Autumn",
                "Winter"
              ]
            }
          },
          {
            "name": "style",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "active",
                "archived",
                "all"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClosetItem"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          },
          "422": {
            "description": "Missing or archived outfit member",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createClosetItem",
        "tags": [
          "closet"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClosetItemInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetItem"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          },
          "422": {
            "description": "Missing or archived outfit member",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          }
        }
      }
    },
    "/wardrobe/{id}": {
      "get": {
        "operationId": "getClosetItem",
        "tags": [
          "closet"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetItem"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          },
          "422": {
            "description": "Missing or archived outfit member",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "updateClosetItem",
        "tags": [
          "closet"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClosetItemPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetItem"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          },
          "422": {
            "description": "Missing or archived outfit member",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "archiveClosetItem",
        "tags": [
          "closet"
        ],
        "description": "Soft archive. Existing outfit references are retained. Restore using PATCH status=active.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetItem"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          },
          "422": {
            "description": "Missing or archived outfit member",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          }
        }
      }
    },
    "/wardrobe/{id}/favorite": {
      "post": {
        "operationId": "toggleClosetItemFavorite",
        "tags": [
          "closet"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetItem"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          },
          "422": {
            "description": "Missing or archived outfit member",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          }
        }
      }
    },
    "/outfits": {
      "get": {
        "operationId": "listClosetOutfits",
        "tags": [
          "closet"
        ],
        "parameters": [
          {
            "name": "favorite",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            }
          },
          {
            "name": "season",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "Spring",
                "Summer",
                "Autumn",
                "Winter"
              ]
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClosetOutfit"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          },
          "422": {
            "description": "Missing or archived outfit member",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createClosetOutfit",
        "tags": [
          "closet"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClosetOutfitInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetOutfit"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          },
          "422": {
            "description": "Missing or archived outfit member",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          }
        }
      }
    },
    "/outfits/{id}": {
      "get": {
        "operationId": "getClosetOutfit",
        "tags": [
          "closet"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetOutfit"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          },
          "422": {
            "description": "Missing or archived outfit member",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "updateClosetOutfit",
        "tags": [
          "closet"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClosetOutfitPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetOutfit"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          },
          "422": {
            "description": "Missing or archived outfit member",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteClosetOutfit",
        "tags": [
          "closet"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          },
          "422": {
            "description": "Missing or archived outfit member",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          }
        }
      }
    },
    "/outfits/{id}/favorite": {
      "post": {
        "operationId": "toggleClosetOutfitFavorite",
        "tags": [
          "closet"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetOutfit"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          },
          "422": {
            "description": "Missing or archived outfit member",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          }
        }
      }
    },
    "/style-profile": {
      "get": {
        "operationId": "getClosetProfile",
        "tags": [
          "closet"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetProfile"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          },
          "422": {
            "description": "Missing or archived outfit member",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "updateClosetProfile",
        "tags": [
          "closet"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClosetProfilePatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetProfile"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          },
          "422": {
            "description": "Missing or archived outfit member",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosetError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ClosetItemPatch": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "description": {
            "type": "string",
            "maxLength": 5000
          },
          "brand": {
            "type": "string",
            "maxLength": 200
          },
          "category": {
            "type": "string",
            "enum": [
              "Tops",
              "Bottoms",
              "Outerwear",
              "Dresses",
              "Shoes",
              "Bags",
              "Accessories",
              "Jewelry"
            ]
          },
          "subcategory": {
            "type": "string",
            "maxLength": 200
          },
          "primaryColor": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "secondaryColors": {
            "maxItems": 50,
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200
            }
          },
          "pattern": {
            "type": "string",
            "maxLength": 200
          },
          "material": {
            "type": "string",
            "maxLength": 200
          },
          "fit": {
            "type": "string",
            "maxLength": 200
          },
          "size": {
            "type": "string",
            "maxLength": 200
          },
          "styleTags": {
            "maxItems": 50,
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200
            }
          },
          "formality": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "Casual",
                  "Smart Casual",
                  "Formal",
                  "Lounge"
                ]
              },
              {
                "type": "null"
              }
            ]
          },
          "seasons": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "Spring",
                "Summer",
                "Autumn",
                "Winter"
              ]
            }
          },
          "occasions": {
            "maxItems": 50,
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200
            }
          },
          "weatherSuitability": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "hot",
                "warm",
                "cool",
                "cold",
                "rain",
                "snow",
                "wind"
              ]
            }
          },
          "originalImage": {
            "type": "string",
            "maxLength": 500
          },
          "processedImage": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ]
          },
          "thumbnail": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ]
          },
          "favorite": {
            "type": "boolean"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "archived"
            ]
          },
          "availability": {
            "type": "string",
            "enum": [
              "available",
              "unavailable"
            ]
          },
          "maintenanceState": {
            "type": "string",
            "enum": [
              "clean",
              "laundry",
              "repair"
            ]
          },
          "wearCount": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "lastWornAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
              },
              {
                "type": "null"
              }
            ]
          },
          "aiGeneratedTags": {
            "maxItems": 50,
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200
            }
          },
          "aiConfidence": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0,
                "maximum": 1
              },
              {
                "type": "null"
              }
            ]
          },
          "aiAnalysisVersion": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ]
          },
          "userCorrectedFields": {
            "maxItems": 50,
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200
            }
          }
        },
        "additionalProperties": false
      },
...
      "ClosetError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "issues": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        },
        "required": [
          "error"
        ]
      }
    }
  }
}